vType: swap {restr}Frame{,2}; coercePaintingType reqd
Converting everything to the first definition lets us make progress
throughout, but Painting fundamentally does not compute on the first
definition, making it impossible to define RestrPainting. We need to
phrase RestrPainting in terms of the first definition, but convert to
the CohFrameTypesAndRestrFrames definition when doing the proof. The
first step is to define coercePaintingType, for which we start to make
progress.